backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.0
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0,<5
python-multipart>=0.0.9
aiosmtplib>=3.0.0
boto3>=1.35.0
pytest>=8.0.0
httpx>=0.27.0
